[feature]improve typesense source

+++ seatunnel-connectors-v2/connector-typesense/src/main/java/org/apache/seatunnel/connectors/seatunnel/typesense/source/TypesenseSourceReader.java
@@ -92,7 +92,7 @@ public class TypesenseSourceReader implements SourceReader<SeaTunnelRow, Typesen
                                 deserializer.deserialize(new TypesenseRecord(document));
                         output.collect(seaTunnelRow);
                     }
-                    if (found / pageSize - 1 > sourceCollectionInfo.getOffset()/pageSize) {
+                    if ((double)found / pageSize - 1 > sourceCollectionInfo.getOffset()/pageSize) {
                         sourceCollectionInfo.setOffset(sourceCollectionInfo.getOffset() + pageSize);
                     }else{
                         break;
